!979 Fix misc bugs * fix misc compile error * tests: Fix shape/offset for distributed tests * Fix misc bugs * Change offset/shape type to int64_t * include: Change shape/offset type to int64 * operator: Change shape/offset type to int64_t * interface: Change Shape/Offset type to int64_t * tests: Change shape type to int64_t for st test * tests: Change shape type to int64_t for ut * pass: Change shape/offset type to std::vector<int64_t> * tensor: Add custom shape for Shape/Offset/Stride

+++ include/tilefwk/symbolic_scalar.h
@@ -137,8 +137,8 @@ public:
     friend std::ostream &operator<<(std::ostream &os, const SymbolicScalar &val) { return os << val.Dump(); }
 
 public:
-    static std::vector<int> Concrete(const std::vector<SymbolicScalar> &scalarList, int64_t defValue);
-    static std::vector<SymbolicScalar> FromConcrete(const std::vector<int> &values);
+    static std::vector<int64_t> Concrete(const std::vector<SymbolicScalar> &scalarList, int64_t defValue);
+    static std::vector<SymbolicScalar> FromConcrete(const std::vector<int64_t> &values);
 
 public:
     /* internal use */
